requirements.txt
requests
BeautifulSoup4
lxml
pydantic
fuzzywuzzy
aiohttp
python-Levenshtein
nest_asyncio
aiofiles
pyyaml
